pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "china-car-parts-bot"
version = "1.0.0"
description = "Chinese Auto Parts Price Bot with Telegram interface"
authors = [
    {name = "China Car Parts Team", email = "admin@example.com"}
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.10.0",
    "aiogram>=3.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "rapidfuzz>=3.0.0",
    "httpx>=0.24.0",
    "requests>=2.31.0"
]

[tool.basedpyright]
include = ["app"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.*"
]
typeCheckingMode = "basic"
pythonVersion = "3.11"
pythonPlatform = "Darwin"
reportMissingImports = true
reportMissingTypeStubs = false
reportImportCycles = true
reportUnusedImport = true
reportUnusedVariable = true
reportDuplicateImport = true
reportOptionalSubscript = false
reportOptionalMemberAccess = false
reportOptionalCall = false
reportOptionalIterable = false
reportOptionalContextManager = false
reportOptionalOperand = false
reportTypedDictNotRequiredAccess = false

[tool.pylint]
load-plugins = ["pylint.extensions.no_self_use"]
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-locals",
    "too-many-branches",
    "too-many-statements",
    "import-error"
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "aiogram.*",
    "fastapi.*",
    "uvicorn.*",
    "sqlalchemy.*",
    "alembic.*"
]
ignore_missing_imports = true
